Addressing code smells
