+++ src/IkeMtz.NRSRx.Core.Web/CoreWebStartup.cs
@@ -4,16 +4,17 @@ using System.Linq;
 using System.Reflection;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
+using Swashbuckle.AspNetCore.SwaggerUI;
 
 namespace IkeMtz.NRSRx.Core.Web
 {
@@ -83,6 +84,59 @@ namespace IkeMtz.NRSRx.Core.Web
        {
          _ = webBuilder.UseStartup<TStartup>();
        });
+    }
+
+    public virtual void SetupSwaggerUI(SwaggerUIOptions options, IApiVersionDescriptionProvider provider)
+    {
+      foreach (var description in provider.ApiVersionDescriptions)
+      {
+        options.SwaggerEndpoint($"./swagger/{description.GroupName}/swagger.json", description.GroupName.ToUpperInvariant());
+      }
+      options.EnableDeepLinking();
+      options.EnableFilter();
+      options.RoutePrefix = string.Empty;
+      options.HeadContent += "<meta name=\"robots\" content=\"none\" />";
+      options.OAuthClientId(Configuration.GetValue<string>("SwaggerClientId"));
+      options.OAuthAppName(Configuration.GetValue<string>("SwaggerAppName"));
+    }
+
+    public virtual void SetupSwaggerGen(SwaggerGenOptions options)
+    {
+      options.UseInlineDefinitionsForEnums();
+      // add a custom operation filter which sets default values
+      options.OperationFilter<SwaggerDefaultValues>();
+      var audiences = GetIdentityAudiences();
+      var swaggerIdentityProviderUrl = Configuration.GetValue<string>("SwaggerIdentityProviderUrl");
+      if (audiences.Any() && !string.IsNullOrWhiteSpace(swaggerIdentityProviderUrl))
+      {
+        var audience = audiences.FirstOrDefault();
+
+        options.AddSecurityDefinition(JwtBearerDefaults.AuthenticationScheme, new OpenApiSecurityScheme
+        {
+          Type = SecuritySchemeType.OAuth2,
+          In = ParameterLocation.Header,
+          Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
+          Scheme = JwtBearerDefaults.AuthenticationScheme,
+          Flows = new OpenApiOAuthFlows
+          {
+            Implicit = new OpenApiOAuthFlow
+            {
+              AuthorizationUrl = new Uri($"{swaggerIdentityProviderUrl}authorize?audience={audience}"),
+              Scopes = SwaggerScopes,
+            },
+          }
+        });
+        options.AddSecurityRequirement(new OpenApiSecurityRequirement
+                {
+                    {
+                        new OpenApiSecurityScheme
+                        {
+                            Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = JwtBearerDefaults.AuthenticationScheme}
+                        },
+                       Array.Empty<string>()
+                    }
+                });
+      }
     }
   }
 }
