+++ src/IkeMtz.NRSRx.Core.WebApi/CoreWebApiStartup.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Linq;
 using IkeMtz.NRSRx.Core.Web;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -12,7 +9,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
-using Microsoft.OpenApi.Models;
 using Newtonsoft.Json;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using WebApiContrib.Core.Formatter.MessagePack;
@@ -55,20 +51,7 @@ namespace IkeMtz.NRSRx.Core.WebApi
        .UseAuthentication()
        .UseAuthorization()
        .UseSwagger()
-       .UseSwaggerUI(options =>
-       {
-         // build a swagger endpoint for each discovered API version
-         foreach (var description in provider.ApiVersionDescriptions)
-         {
-           options.SwaggerEndpoint($"./swagger/{description.GroupName}/swagger.json", description.GroupName.ToUpperInvariant());
-         }
-         options.EnableDeepLinking();
-         options.EnableFilter();
-         options.RoutePrefix = string.Empty;
-         options.HeadContent += "<meta name=\"robots\" content=\"none\" />";
-         options.OAuthClientId(Configuration.GetValue<string>("SwaggerClientId"));
-         options.OAuthAppName(Configuration.GetValue<string>("SwaggerAppName"));
-       })
+       .UseSwaggerUI(options => SetupSwaggerUI(options, provider))
        .UseEndpoints(endpoints =>
       {
         _ = endpoints.MapControllers();
@@ -122,40 +105,7 @@ namespace IkeMtz.NRSRx.Core.WebApi
         .AddSwaggerGen(options =>
         {
           options.UseInlineDefinitionsForEnums();
-          // add a custom operation filter which sets default values
-          options.OperationFilter<SwaggerDefaultValues>();
-          var audiences = GetIdentityAudiences();
-          var swaggerIdentityProviderUrl = Configuration.GetValue<string>("SwaggerIdentityProviderUrl");
-          if (audiences.Any() && !string.IsNullOrWhiteSpace(swaggerIdentityProviderUrl))
-          {
-            var audience = audiences.FirstOrDefault();
-
-            options.AddSecurityDefinition(JwtBearerDefaults.AuthenticationScheme, new OpenApiSecurityScheme
-            {
-              Type = SecuritySchemeType.OAuth2,
-              In = ParameterLocation.Header,
-              Description = "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
-              Scheme = JwtBearerDefaults.AuthenticationScheme,
-              Flows = new OpenApiOAuthFlows
-              {
-                Implicit = new OpenApiOAuthFlow
-                {
-                  AuthorizationUrl = new Uri($"{swaggerIdentityProviderUrl}authorize?audience={audience}"),
-                  Scopes = SwaggerScopes,
-                },
-              }
-            });
-            options.AddSecurityRequirement(new OpenApiSecurityRequirement
-                {
-                    {
-                        new OpenApiSecurityScheme
-                        {
-                            Reference = new OpenApiReference { Type = ReferenceType.SecurityScheme, Id = JwtBearerDefaults.AuthenticationScheme}
-                        },
-                       Array.Empty<string>()
-                    }
-                });
-          }
+          SetupSwaggerGen(options);
         });
     }
